Feature language switcher

+++ wp-content/plugins/sitepress-multilingual-cms/classes/taxonomy-term-translation/class-wpml-term-actions.php
@@ -37,24 +37,29 @@ class WPML_Term_Actions extends WPML_Full_Translation_API {
 	 */
 	function delete_term_actions( $tt_id, $taxonomy ) {
 		$icl_el_type = 'tax_' . $taxonomy;
-
-		if ( $this->sitepress->get_setting( 'sync_delete_tax' ) && $this->delete_recursion_flag === false ) {
-			// only for translated
-			$lang_details = $this->sitepress->get_element_language_details( $tt_id, $icl_el_type );
-			if ( empty( $lang_details->source_language_code ) ) {
-				// get translations
-				$trid                        = $this->sitepress->get_element_trid( $tt_id, $icl_el_type );
-				$translations                = $this->sitepress->get_element_translations( $trid, $icl_el_type );
-				$this->delete_recursion_flag = true;
-				// delete translations
-				remove_filter( 'get_term', array( $this->sitepress, 'get_term_adjust_id' ), 1 );
-				foreach ( $translations as $translation ) {
-					if ( (int) $translation->element_id !== (int) $tt_id ) {
-						wp_delete_term( $translation->term_id, $taxonomy );
-					}
+		$trid        = $this->sitepress->get_element_trid( $tt_id, $icl_el_type );
+		$lang_details = $this->sitepress->get_element_language_details( $tt_id, $icl_el_type );
+		if ( $this->sitepress->get_setting( 'sync_delete_tax' )
+		     && $this->delete_recursion_flag === false
+		     && empty( $lang_details->source_language_code )
+		) {
+			// get translations
+			$translations                = $this->sitepress->get_element_translations( $trid, $icl_el_type );
+			$this->delete_recursion_flag = true;
+			// delete translations
+			$has_filter = remove_filter( 'get_term', array( $this->sitepress, 'get_term_adjust_id' ), 1 );
+			foreach ( $translations as $translation ) {
+				if ( (int) $translation->element_id !== (int) $tt_id ) {
+					wp_delete_term( $translation->term_id, $taxonomy );
 				}
+			}
+			if ( $has_filter ) {
 				add_filter( 'get_term', array( $this->sitepress, 'get_term_adjust_id' ), 1, 1 );
-				$this->delete_recursion_flag = false;
+			}
+			$this->delete_recursion_flag = false;
+		} else {
+			if ( empty( $lang_details->source_language_code ) ) {
+				$this->set_new_original_term( $trid, $lang_details->language_code );
 			}
 		}
 
@@ -67,8 +72,40 @@ class WPML_Term_Actions extends WPML_Full_Translation_API {
 		do_action( 'wpml_translation_update', array_merge( $update_args, array( 'type' => 'before_delete' ) ) );
 
 		$this->wpdb->delete( $this->wpdb->prefix . 'icl_translations', array( 'element_type' => $icl_el_type, 'element_id' => $tt_id ) );
-
 		do_action( 'wpml_translation_update', array_merge( $update_args, array( 'type' => 'after_delete' ) ) );
+	}
+
+	/**
+	 * @param $trid
+	 * @param $deleted_language_code
+	 */
+	public function set_new_original_term( $trid, $deleted_language_code ) {
+		if ( $trid && $deleted_language_code ) {
+			$order_languages = $this->sitepress->get_setting( 'languages_order' );
+			$this->term_translations->reload();
+			$translations         = $this->term_translations->get_element_translations( false, $trid );
+			$new_source_lang_code = false;
+			foreach ( $order_languages as $lang_code ) {
+				if ( isset( $translations[ $lang_code ] ) ) {
+					$new_source_lang_code = $lang_code;
+					break;
+				}
+			}
+			if ( $new_source_lang_code ) {
+				$rows_updated = $this->wpdb->update( $this->wpdb->prefix . 'icl_translations',
+					array( 'source_language_code' => $new_source_lang_code ),
+					array( 'trid' => $trid, 'source_language_code' => $deleted_language_code )
+				);
+
+				if ( 0 < $rows_updated ) {
+					do_action( 'wpml_translation_update', array( 'trid' => $trid ) );
+				}
+
+				$this->wpdb->query( "UPDATE {$this->wpdb->prefix}icl_translations 
+									 SET source_language_code = NULL 
+									 WHERE language_code = source_language_code" );
+			}
+		}
 	}
 
 	/**
